Allow project to use aphanumeric character as id.
This is a huge patch. It removes completely the project parameters and
the id, which was part of, and rewrite some stuff of configfile which
were tightly coupled on project parameters. Bricks had lost
project_parms too.

Now when loading a project is responsibility of the caller to reset the
factory state and the filename is used also as base for the private
images.