Fix a possible race problem with *symm attributes.
If *symm is set to False and at same time right-to-left parameter is
changed, it could be possible that the *symm parameter is set before
and the successive parameter set cause an asymmetry in the parameters.

i.e.

losssymm = False
loss = 10
lossr = 3

if *both* losssymm and lossr are set respectively to True and 5 the
order of commands is not certain the could be sent as follow:

	loss 10
	loss RL 5

causing the asymmetry.